Generic clustering function

Generic clustering function for different object classes.

Function to calculate and cluster on the skew-geometric Jensen-Shannon Divergence for the conditional extremes model. When comparing conditional extremes fits for a single variable using the JS-G between any two locations, we need to use the same "data" for each location. Therefore, we look at values from the maximum of the thresholds at each location, to some multiple dat_max_mult of this, with n_dat equally spaced points. cluster:pam is used to cluster the Jensen-Shannon divergence distance matrix between all locations, summed across all variables.

Function to cluster on the skew-geometric Jensen-Shannon Divergence distance matrix for the conditional extremes model.

cecl_clust(x, ...)

# S3 method for class 'cecl_dep'
cecl_clust(
  x,
  marg_obj,
  k,
  var = NULL,
  cluster_mem = NULL,
  laplace_cap = 0.99,
  n_mc = 500,
  ncores = 1,
  par_dist = FALSE,
  seed = NULL,
  ...
)

# S3 method for class 'cecl_dist'
cecl_clust(x, k, var = NULL, cluster_mem = NULL, ...)

Arguments

x

Object of class cecl_dist.

...

Additional arguments passed to methods.

marg_obj

Object of class cecl_marg.

k

Number of clusters.

var

Optional conditioning variable name to cluster on, if not all variables, Default: NULL.

cluster_mem

Optional vector of true cluster memberships to evaluate clustering solution, Default: NULL.

laplace_cap

Upper quantile to sample Laplace distribution truncation point from, Default: 0.99.

n_mc

Number of Monte Carlo samples to use in distance calculation, Default: 500.

ncores

Number of cores to use for parallel computation, Default: 1.

par_dist

Logical, whether to parallelise the distance computation, rather than embarrassingly parallel over each variable, which is useful for a high number of locations, but has significant overheads, Default: FALSE.

seed

Seed number for random number generation in Monte Carlo sampling, for reproducibility.

Value

Clustering results.

List containing the clustering results and, if cluster_mem is provided, the adjusted Rand index.

List containing the clustering results and, if cluster_mem is provided, the adjusted Rand index.