Fit marginal models for multivariate conditional extremes

Fit marginal models for multivariate conditional extremes.

cecl_marg(
  data,
  vars = NULL,
  thresh_method = c("value", "quantile", "regression", "none"),
  thresh_args = NULL,
  thresh_only = FALSE,
  marg_method = c("ecdf", "ismev", "evgam"),
  marg_args = NULL,
  ncores = 1
)

Arguments

data

List of dataframes at each location containing data for each variable. Optionally, can be an object of class evc_marg from a previous call to fit_ce, with elements marginal, data_thresh and original.

vars

Names of variable columns.

thresh_method

Method to select marginal thresholds, one of "value" (fixed value), "quantile" (quantile thresholding), "regression" (covariate dependent thresholding via qgam) or "none", for no thresholding (or where you have already thresholded the data prior to input).

thresh_args

Arguments to be passed to thresholding function. For "value", a numeric value (scalar for shared value or vector for each vars) for simple value thresholding. For "quantile", a numeric value (scalar for shared value or vector for each vars) for quantile thresholding. For "regression", a list of arguments to be passed to qgam_thresh (see documentation for details).

thresh_only

If TRUE, only threshold data and do not fit marginal models, Default: FALSE.

marg_method

Method to fit marginal models, one of "ecdf" (empirical CDF), "ismev" (using ismev::gpd.fit) or "evgam" (using evgam::evgam).

marg_args

Arguments to be passed to marginal fitting function.

ncores

Number of cores to use for parallel computation, Default: 1.

Value

Object of type cecl_marg for each location.