Threshold data for cecl_marg

Threshold data for cecl_marg using specified method.

marg_thresh(
  data_df,
  mult_col = "name",
  vars,
  thresh_method,
  thresh_args,
  ret_obj = TRUE
)

Arguments

mult_col

Name of column specifying multivariate structure, default "name".

vars

Names of variable columns.

thresh_method

Method to select marginal thresholds, one of "value" (fixed value), "quantile" (quantile thresholding), "qgam" (covariate dependent thresholding via qgam) or "none", for no thresholding (or where you have already thresholded the data prior to input).

thresh_args

Arguments to be passed to thresholding function. For "value", a numeric value (scalar for shared value or vector for each vars) for simple value thresholding. For "quantile", a numeric value (scalar for shared value or vector for each vars) for quantile thresholding. For "qgam", a list of arguments to be passed to qgam_thresh (see documentation for details).

ret_obj

If TRUE and thresh_method is "qgam" or

marg_method is "ismev" or "evgam", return fitted marginal models. Default: TRUE.

Value

List of thresholded data frames for each variable.