Fit marginal models for multivariate conditional extremes

Fit marginal models for multivariate conditional extremes.

cecl_marg(
  data,
  mult_col = "name",
  vars = NULL,
  thresh_method = c("value", "quantile", "qgam", "none"),
  thresh_args = NULL,
  thresh_only = FALSE,
  marg_method = c("ecdf", "ismev", "evgam"),
  marg_args = NULL,
  ret_obj = TRUE,
  ncores = 1
)

Arguments

data

List of dataframes at each location containing data for each variable. Optionally, can be an object of class evc_marg from a previous call to fit_ce, with elements marginal, data_thresh and original.

mult_col

Name of column specifying multivariate structure, default "name".

vars

Names of variable columns.

thresh_method

Method to select marginal thresholds, one of "value" (fixed value), "quantile" (quantile thresholding), "qgam" (covariate dependent thresholding via qgam) or "none", for no thresholding (or where you have already thresholded the data prior to input).

thresh_args

Arguments to be passed to thresholding function. For "value", a numeric value (scalar for shared value or vector for each vars) for simple value thresholding. For "quantile", a numeric value (scalar for shared value or vector for each vars) for quantile thresholding. For "qgam", a list of arguments to be passed to qgam_thresh (see documentation for details).

thresh_only

If TRUE, only threshold data and do not fit marginal models, Default: FALSE.

marg_method

Method to fit marginal models, one of "ecdf" (empirical CDF), "ismev" (using ismev::gpd.fit) or "evgam" (using evgam::evgam).

marg_args

Arguments to be passed to marginal fitting function.

ret_obj

If TRUE and thresh_method is "qgam" or marg_method is "ismev" or "evgam", return fitted marginal models. Default: TRUE.

ncores

Number of cores to use for parallel computation, Default: 1.

Value

Object of type cecl_marg for each location.

Examples

# simulate some data
set.seed(123)
n_locs <- 5
n <- 1000
vars <- c("X1", "X2")
df <- do.call(
  rbind,
  lapply(1:n_locs, function(i) {
    data.frame(
      X1 = rgpd(n, u = 0, sigma = 1, xi = 0.2),
      X2 = rgpd(n, u = 0, sigma = 2, xi = -0.1),
      name = paste0("loc_", i)
    )
  })
)
# fit marginal models with quantile thresholding and ISMEV GPD fits
marg_fit <- cecl_marg(
  df,
  thresh_method = "quantile",
  thresh_args = 0.9,
  marg_method = "ismev",
  ncores = 1
)
# inspect output
marg_fit
#> Call:
#> cecl_marg(data = df, thresh_method = "quantile", thresh_args = 0.9, 
#>     marg_method = "ismev", ncores = 1)
#> 
#> Number of locations: 5 
#> Variables: X1, X2 
#> Marginal method: ismev 
summary(marg_fit)
#> Call:
#> cecl_marg(data = df, thresh_method = "quantile", thresh_args = 0.9, 
#>     marg_method = "ismev", ncores = 1)
#> 
#> Number of locations: 5 
#> Variables: X1, X2 
#> Marginal method: ismev 
#> 
#> Marginal Model Coefficients:
#>     name var   thresh    sigma          xi
#> 1  loc_1  X1 2.769196 1.684660  0.10888245
#> 2  loc_1  X2 4.076033 2.252270 -0.27407392
#> 3  loc_2  X1 2.769196 1.272277  0.33943335
#> 4  loc_2  X2 4.076033 1.449344 -0.09515565
#> 5  loc_3  X1 2.769196 1.370116  0.31656787
#> 6  loc_3  X2 4.076033 1.613638 -0.12730872
#> 7  loc_4  X1 2.769196 2.236876 -0.06918687
#> 8  loc_4  X2 4.076033 1.666727 -0.23213904
#> 9  loc_5  X1 2.769196 1.581356  0.30730479
#> 10 loc_5  X2 4.076033 1.216154  0.11183298

# fit marginal models using evgam
marg_fit_evgam <- cecl_marg(
  df,
  thresh_method = "quantile",
  thresh_args = 0.9,
  marg_method = "evgam",
  marg_args = list(f = list("excess ~ name", "~ name")),
  ncores = 1
)
#> Joining with `by = join_by(name)`
#> Joining with `by = join_by(name)`

marg_fit_evgam
#> Call:
#> cecl_marg(data = df, thresh_method = "quantile", thresh_args = 0.9, 
#>     marg_method = "evgam", marg_args = list(f = list("excess ~ name", 
#>         "~ name")), ncores = 1)
#> 
#> Number of locations: 5 
#> Variables: X1, X2 
#> Marginal method: evgam